refactor: components 폴더구조 수정
